requirements.txt
python-amazon-paapi>=6.2.0
requests
rakuten-ws
google-api-python-client
pytrends
feedparser
python-frontmatter
beautifulsoup4
